+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 ## 1.1.4
 * Add optional id to Flowlog.
+* Change some return-value to no-null.
 ## 1.1.3
 * Fix i18n bug.
 ## 1.1.2